addressed Windows warnings from NrrdIO
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@2030 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/nrrd/keyvalue.c
@@ -185,7 +185,7 @@ nrrdKeyValueGet(const Nrrd *nrrd, const char *key) {
 
 void
 _nrrdFwriteEscaped(FILE *file, const char *str) {
-  int ci;
+  size_t ci;
 
   for (ci=0; ci<strlen(str); ci++) {
     switch(str[ci]) {
